Create InvitationService
